+++ changelog.md
@@ -2,7 +2,7 @@
 (because I love changelogs) <br>
 Only changes I consider notable are recorded here, no bugfixes or small changes.
 ### Most Recent
-**v0.0.11c** In-game Changelog Integration. <br>
+**v0.0.12** Shop unlock systems, Upgrades 1-4, and Upgrade ⬥ <br>
 ### Pre-Alpha
 ***
 **v0.0.1** Created this repo, created the base files, added Saving, Loading, Import, and Export systems, and added a favicon. <br>
@@ -19,3 +19,4 @@ Only changes I consider notable are recorded here, no bugfixes or small changes.
 **v0.0.11** Major UI Update. <br>
 **v0.0.11b** Improved Full Reset. <br>
 **v0.0.11c** In-game Changelog Integration. <br>
+**v0.0.12** Shop unlock systems, Upgrades 1-4, and Upgrade ⬥ <br>